Less warning for ChunkStack.cpp

+++ src/ChunkStack.cpp
@@ -109,7 +109,7 @@ void ChunkStack::Collapse()
 {
    size_t wr_idx = 0;
 
-   for (size_t rd_idx = 0; rd_idx < (int)m_cse.size(); rd_idx++)
+   for (size_t rd_idx = 0; rd_idx < m_cse.size(); rd_idx++)
    {
       if (m_cse[rd_idx].m_pc != NULL)
       {
